requirements_ehr.txt
# JEPA-EHR Requirements
# Core dependencies for JEPA-EHR training

# PyTorch (install appropriate version for your CUDA)
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Configuration
pyyaml>=6.0

# Optional but recommended
matplotlib>=3.4.0  # For visualization
scikit-learn>=1.0.0  # For downstream tasks and evaluation
tensorboard>=2.10.0  # For monitoring (optional)

# For MIMIC-IV preprocessing (if needed)
# pyarrow>=10.0.0  # For parquet files
# sqlalchemy>=1.4.0  # For database access
